125483 add component PluginsExtensionPageContents
