+++ apps/app/src/components/Layout/AdminLayout.tsx
@@ -13,8 +13,6 @@ import styles from './Admin.module.scss';
 const PageCreateModal = dynamic(() => import('../PageCreateModal'), { ssr: false });
 const SystemVersion = dynamic(() => import('../SystemVersion'), { ssr: false });
 const HotkeysManager = dynamic(() => import('../Hotkeys/HotkeysManager'), { ssr: false });
-const PluginDeleteModal = dynamic(() => import('~/features/growi-plugin/client/components/Admin/PluginsExtensionPageContents/PluginDeleteModal')
-  .then(mod => mod.PluginDeleteModal), { ssr: false });
 
 
 type Props = {
@@ -49,7 +47,6 @@ const AdminLayout = ({
         </div>
 
         <PageCreateModal />
-        <PluginDeleteModal />
         <SystemVersion />
       </div>
